requirements.txt
streamlit
pandas
matplotlib
numpy
plotly
fpdf
fpdf
requests
pillow